Tabula rasa.
I've always been a fan of the Unix philosophy, and have tried to
apply its lessons in full-stack Web development. This has been
considerably empowering, but it has become quite clear to me that
the industry is trying to move in a somewhat different direction
right now.

While I'm waiting for them to come to their senses, and for myself
to regain enough executive function to approach my old goal of
building multimedia software, I'm gonna try my hand at this
very, very, *very* simple command-line utility.

It simply keeps track of when was the last time you did something,
and how often you do it on average. My goals with this project are:

- First and foremost, to integrate it in my daily life to have some
  visibility over my habits, both good and bad.

- To build something small enough and simple enough to be useful
  to many people in general scenarios.

For now, I've scaffolded a basic Rust project with `clap` and
`microxdg`, which parses command-line options and establishes
state locations.

I've also defined some example usage in the README and will be
working to implement that.